# Web App Project/sportspro/server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SportsPro API",
    "version": "1.0.0",
    "description": "A simple API to perform CRUD functions for sports management software"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/",
      "description": "Version 1"
    }
  ],
  "paths": {
...
    "/incidents/assigned": {
      "get": {
        "summary": "Get incidents assigned to the current user",
        "tags": [
          "Incidents"
        ],
        "responses": {
          "200": {
            "description": "List of assigned incidents"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/incidents/unassigned": {
      "get": {
        "summary": "Get all unassigned incidents",
        "tags": [
          "Incidents"
        ],
        "responses": {
          "200": {
            "description": "List of unassigned incidents"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/incidents": {
      "get": {
        "summary": "Get all incidents",
        "tags": [
          "Incidents"
        ],
        "responses": {
          "200": {
            "description": "List of all incidents"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new incident",
        "tags": [
          "Incidents"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customerid",
                  "techid",
                  "productcode",
                  "dateopened",
                  "title",
                  "description"
                ],
                "properties": {
                  "customerid": {
                    "type": "integer"
                  },
                  "techid": {
                    "type": "integer"
                  },
                  "productcode": {
                    "type": "string"
                  },
                  "dateopened": {
                    "type": "string",
                    "format": "date"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Incident created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/incidents/{incidentid}": {
      "get": {
        "summary": "Get an incident by ID",
        "tags": [
          "Incidents"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentid",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the incident"
          }
        ],
        "responses": {
          "200": {
            "description": "Incident data"
          },
          "404": {
            "description": "Incident not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing incident",
        "tags": [
          "Incidents"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "incidentid",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the incident"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customerid",
                  "techid",
                  "productcode",
                  "dateopened",
                  "title",
                  "description"
                ],
                "properties": {
                  "customerid": {
                    "type": "integer"
                  },
                  "techid": {
                    "type": "integer"
                  },
                  "productcode": {
                    "type": "string"
                  },
                  "dateopened": {
                    "type": "string",
                    "format": "date"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Incident updated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Incident not found"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "Get all products",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of products retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productcode",
                  "name",
                  "version",
                  "releasedate"
                ],
                "properties": {
                  "productcode": {
                    "type": "string",
                    "example": "ABC123"
                  },
                  "name": {
                    "type": "string",
                    "example": "WidgetX"
                  },
                  "version": {
                    "type": "string",
                    "example": "1.0.0"
                  },
                  "releasedate": {
                    "type": "string",
                    "example": "2024-01-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/products/{productcode}": {
      "get": {
        "summary": "Get a product by code",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product code"
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a product by code",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product code"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
...
}